Fix stop-the-world race.
With the associated commit in the mmtk-openjdk repository,
stop-the-world and start-the-world requests no longer have to be made in
the MMTk coordinator thread.

Considering that some VM may still require the thread that stops the
world and the thread that starts the world to be the same, we added a
constant in the Collection trait so that VM bindings can override it
when needed.