Set SFT for MallocSpace (#476)
Set SFT entries for MallocSpace (closes #473), and use the newly added `is_mmtk_object` SFT method to correctly check whether an object in a MallocSpace chunk is allocated by `malloc`.

Fix an issue that if the address is chunk aligned, the SFT entries may not be set properly (closes #374). The `update` method of SFT accepts the number of bytes as the size argument, rather than the number of chunks.