Introduce context for the GC controller thread
This commit introduces the `GCController` struct.  It is owned by the GC
controller thread.  It owns the `GCWorker` for the controller, and has
references to objects that it needs to communicate with other threads.

Previously, those dependencies were scattered in
`ControllerCollectorContext` and `GCWorkScheduler`, which complicated
the initialization of the MMTk instance and the GC controller thread.