Introduce a context object for GC controller thread (#533)
* Introduce context for the GC controller thread

This commit introduces the `GCController` struct.  It is owned by the GC
controller thread.  It owns the `GCWorker` for the controller, and has
references to objects that it needs to communicate with other threads.

Previously, those dependencies were scattered in
`ControllerCollectorContext` and `GCWorkScheduler`, which complicated
the initialization of the MMTk instance and the GC controller thread.

* Move controller functions to GCController

Move the GCWorkScheduler::wait_for_completion method to the GCController
type.  This allows us to move controller-specific fields away from
GCWorkScheduler, too.

* Miscellaneous fixes.

-   Renamed ControllerCollectorContext to GCRequester
-   Removed initializer and finalizer functions from GCWorkScheduler
-   The embedded GCController is no longer boxed.
-   Added comments
-   Minor adjustment to API functions.
-   Other minor fixes.