Split the TransitiveClosure trait (#566)
* Split the TransitiveClosure trait

Moved the `process_edge` function of the `TransitiveClosure` trait into
a dedicated trait `EdgeVisitor` for visiting edges of an object (or
stack roots).  As a result,

1.  VM bindings only require the `EdgeVisitor` trait for object
    scanning.  VM bindings no longer use the `TransitiveClosure` in any
    way.

2.  Now `TransitiveClosure` only serves one purpose in `mmtk-core`, i.e.
    as a super-trait of ProcessEdgesWork for the `process_node` method.
    It is only called by `trace_object` of different spaces.

With this change, it will be easy to refactor `trace_object` and remove
`TransitiveClosure` completely.

* Fix scan_object{,s} API

Now both scan_object and scan_objects take EdgeVisitor as parameter.

We now implement scan_objects in mmtk-core.  We no longer expect the
VM bindings to create ProcessEdgesWork in scan_objects.