Decouple root scanning from ProcessEdgesWork.
During root scanning, the VM should only report lists of roots.  It is
the responsibility of mmtk-core (particularly the Plan), not the VM, to
decide what work packet to create for the roots.

This commit introduces a RootsHandlerFactory.  It provides factory
methods for creating work packets from a list of edges (for updating
root edges) or nodes (which pins the roots).  Concrete plans provide
concrete factories and decide what work packets to use.