Replace TransitiveClosure with an ObjectQueue trait (#607)
We replaced the TransitiveClosure parameter of the `trace_object` method
of spaces with ObjectQueue.  Wherever it called `process_node`, it now
calls `ObjectQueue::enqueue`, instead.  By doing this, we decoupled
`trace_object` from ProcessEdgesWork which was the only implementor of
TransitiveClosure.

ObjectQueue is a trait.  This means the data structure that holds object
references and the behaviour of enqueuing objects can be customised,
too.  The default VectorObjectQueue implementation behaves like the old
`ProcessEdgesWork::process_node`.

This commit removes the TransitiveClosure trait.