Minimize the scope of acquire_lock (#614)
This PR changes the scope of `Space.acquire_lock`. The lock was introduced in https://github.com/mmtk/mmtk-core/pull/555, and it caused bad allocation performance when we have many threads (e.g. more than 24 threads, mutator threads or GC threads with copying allocators). This pull request reduces the scope of this `acquire_lock`, and the overhead from it is mostly neglectable. 

This closes https://github.com/mmtk/mmtk-core/issues/610.