Remove Plan.gc_init() and Space.init() (#629)
This pull request removes `Plan.gc_init()`, and `Space.init()`. I think this can close the issue: https://github.com/mmtk/mmtk-core/issues/57, I am not aware of any occurrences of two-step initialisation in MMTk after this PR.

Note that We eagerly set SFT for each space in their `Space.init()` if they are contiguous and SFT needs a non-moving reference to the `Space`. When we create a `Space`, the reference points to an address on the stack, and will be moved. So we cannot set SFT at creation. I added a method `Space.initialize_sft()` which is called after the `Space` has a fixed address. `initialize_sft()` is only used to set SFT, not any other post-creation initialisation.