Feature to turn Immix into non-moving GC (#633)
* Feature to turn Immix into non-moving GC

Added some features to disable defragmentation and make Immix a
non-moving GC.  It is useful during developing new VM bindings so that
the VM can use Immix before the VM fully supports object moving.

The `immix_no_defrag` feature disables defragmentation in ImmixSpace.
The Immix Plan now reports itself as a non-moving collector
(IMMIX_CONSTRAINTS.moves_objects = false) if policy::immix::DEFRAG is
false.

The `immix_smaller_block` feature reduces the size of blocks of
ImmixSpace.  This should mitigate fragmentation when defragmentation is
disabled.

* Work-around metadata and space overlapping

Moved the metadata base address to a higher address so that it is less
likely to clash with spaces.