Add options for setting GC thread affinity (#670)
This commit allows the user to set GC thread affinity by specifying a
list of cores. The cores are specified similar to how a CPU list is
provided to `taskset` on Linux, i.e. with discrete core ids or ranges.

Note that this feature currently only works on Linux due to
platform-specific calls.

Closes #299.