Mask and zero bits in metadata bytes for bulk zeroing (#707)
This PR fixes a bug in `bzero_metadata()` discussed in https://github.com/mmtk/mmtk-core/pull/643#issuecomment-1331875577. When the data address range to be bulk zeroed cannot be mapped into whole bytes in the side metadata, the other bits in the bytes which should not be updated will get zeroed unexpectedly.