Generalise weak reference processing for other languages (#700)
This commit adds a language-neutral API to let the VM binding handle weak
references and finalisers.

Added VMRefClosure and VMRefForwarding work bucket stages to replace the
Java-specific stages.

Added a "sentinel" mechanism to execute a work packet when a bucket is drained,
and prevent the GC from entering the next stage, making it possible to expand
the transitive closure multiple times. It replaces GCWorkScheduler::closure_end.

Extended the Collection::process_weak_refs method to allow the binding to trace
objects during weak reference processing.

Renamed Collection::vm_release to Collection::post_forwarding, and it is now
executed in a dedicated work packet.