Workaround a synchronization bug.
If a worker thread spuriously wakes up, it may observe the pending
coordinator work counter going from 0 to 1 when the EndOfGC work packet
starts to be executed.  That'll trigger an assertion failure.

It is a workaround because a full solution should explicitly address the
need to prevent opening more work packets when the coordinator (or any
worker) is adding work packets to related buckets.