Fix stale forwarding bits bug
When copying, we erroneously assumed that the forwarding bits of the new
copy must not have been set if the forwarding bits metadata is on the
side.  If the new copy is allocated in a previously free chunk, it may
contain stale forwarding bits from the previous GC.

Now we clear forwarding bits in post_copy regardless whether the
forwarding bits metadata is in the header or on the side.

Given that post_copy always clears forwarding bits, PrepareBlockState
now only clears forwarding bits for defrag source blocks of ImmixSpace.

Closes: https://github.com/mmtk/mmtk-core/issues/776