Add C2 slowpath allocation barrier (#785)
Add the missing C2 slow-path allocation barrier.

The C2 slow-path allocation code can do deoptimization after the
allocation and before returning to C2 compiled code. The deoptimization
itself contains a safepoint.

For generational plans, if a GC happens at this safepoint, the allocated
object will be promoted, and all the subsequent field initialization
should be recorded. The solution is to unconditionally treat this object
as being written if the object is not in nursery, and record this object
in the modbuf. This is also the solution for G1:
https://github.com/mmtk/openjdk/blob/e1e420b3c4fd627218e45ec977972561c559acec/src/hotspot/share/gc/shared/cardTableBarrierSet.cpp#L136-L160

[mmtk-openjdk PR](https://github.com/mmtk/mmtk-openjdk/pull/205)

---------

Co-authored-by: Yi Lin <qinsoon@gmail.com>