No coordinator work
Remove the concept of "coordinator work packets".  Now
ScheduleCollection and StopMutators are both executed on ordinary
workers.  The only work packet executed by the coordinator is EndOfGC.

Simplified the interaction between the coordinator and the workers.  The
coordinator only responds to the event that "all workers have parked".
Removed the workers-to-coordinator channel.  WorkerMonitor now has two
Condvars, one (the existing one) for notifying workers about more work
available, and another for notifying the coordinator that all workers
have parked.