Let ObjectReference implement Ord (#870)
This will enable `Vec<ObjectReference>` to be sorted, and enable
`ObjectReference` to be used as the keys of ordered collections such as
`BTreeMap`. Although mmtk-core currently never sorts
`Vec<ObjectReference>`, developers of mmtk-core as well as bindings may
experiment with accessing `ObjectReference` instances from low addresses
to high addresses. It complicates things if `ObjectReference` does not
implement `Ord`.