Collect live bytes during GC (#768)
This PR adds a feature `count_live_bytes_in_gc`. When enabled, each
worker will maintain a counter for live bytes when scanning an object,
and in `Release`, we get live bytes from all the workers, and sum it up.
We also provide `memory_manager::live_bytes_in_last_gc()` for users to
query this value. This is usually used to debug fragmentation.