Implement AllocatorInfo (#889)
This type adds `AllocatorInfo` enum that can be used to query offset of
allocator fields for fast-path generation. Currently all bindings have
to re-declare `*Allocator` types with the same layout as in mmtk-core in
order to get offset of these fields. This enum simplifies that, all you
need to do is invoke `get_allocator_info(selector)`.