No bind_freelist
No bind_freelist.

No having references to FreeList from multiple places.

Enumerate free lists from each space instead from global arrays.

If Map64::allocate_contiguous_chunks wants to update RawMemoryFreeList,
we pass a &mut RawMemoryFreeList to it.

If Map32::finalize_static_space_map wants to update the start addresses
of spaces (now page resources'), we give it a call back so that we can
call a method of the plan to update the starts for it at the right time.