Benchmark Rust code (#933)
This PR allows benchmarking mmtk-core code in DummyVM. It uses
`criterion`, which is a popular framework to benchmark with a stable
Rust toolchain. This PR just introduces code and two benchmarks that
specifically test our allocation code in Rust and our SFT read access.
This PR does not include those benchmarks in the CI.