Annotate PageProtext::common with #[parent]
Added the missing `#[parent]` attribute to the `PageProtect::common`
field.  It didn't have the `#[fallback_trace]` attribute, but it worked
because the PageProtect plan only ever allocates into
`PageProtect::space` which is an LOS.  Now that the `#[parent]`
attribute is also used for enumerating spaces and verifying space
metadata, it can no longer be omitted.  Omitting `#[parent]` will cause
failure when the "extreme_assertions" feature is enabled.