Use atomic operations for SFT map and remove unsafe code (#931)
This PR removes unsafe code from SFT map.
* For most SFT map implementations, we only need mutability for the
cells in the table that stores SFT pointers. We do not need mutability
for the table itself. In this case, using atomic operations solves the
problem.
* For `DenseChunkMap`, we do need to mutate the table. But we only
mutate the table during plan creation. We can use unsafe code to get a
mutable reference to SFT map during plan creation (and it is easy to
reason that getting a mutable reference here is correct).

Changes in the PR
* Add a type `SFTRefStorage` which is internally implemented as a double
word atomic integer (u128 or u64). We store `SFTRefStorage` in SFT maps.
* Repurpose `SFTMap::eager_initialize` to
`SFTMap::initialize_for_space`. It has to be called for each space
created during plan creation time, and it takes a mutable reference to
the `SFTMap` to allow mutating. `start` and `size` is changed to
optional now, as `Option<(Address, usize)>. If the range is given, we
will eagerly initialize SFT map for the region.
* `Space::initialize_sft(&self)` is changed to
`Space::initialize_sft(&self, sft_map: &mut dyn SFTMap)` to allow
mutating SFT map.
* Add `InitializeOnce::get_mut`.
* Bump MSRV to 1.70 due to
https://github.com/mmtk/mmtk-core/pull/931#issuecomment-1700518556.