Fix length of Map64::descriptor_map (#956)
Its length should be MAX_SPACES instead of MAX_CHUNKS.

The size was `MAX_SPACES` in the JikesRVM MMTk. It may have been a typo
when porting JikesRVM MMTk to Rust.

The new `Map64::descriptor_map` does not deserve the unsafe
`new_zeroed_vec` function which was introduced to solve the slow
start-up problem. We keep `new_zeroed_vec` just in case we may still
need it in the future.