Do not reset ImmixAllocator when preparing mutator
It is unnecessary to reset the ImmixAllocator of a mutator when
preparing mutator because the mutator's ImmixAllocator is not used
during GC.  When a worker defragments the heap and promotes young
objects, it uses the ImmixAllocator in ImmixCopyContext or
ImmixHybridCopyContext.

The ImmixAllocator of a mutator still needs to be reset after GC (when
releasing mutators) because the blocks it was allocating into could have
been defragmented or recycled.  So it needs to reset the bump pointers.