Clear all side forwarding bits (#974)
This reverts an optimization which was intended to reduce the number of
side forwarding bits cleared for ImmixSpace by clearing the forwarding
bits metadata of defrag sources, only. This approach does not work for
StickyImmix because it may move young objects, which is not "defrag".

Now ImmixSpace clears the on-the-side forwarding bits for all chunks
just like how it clears the on-the-side marking bits.

This PR prefers correctness over performance, and may degrade the
performance of VMs that use on-the-side forwarding bits. Currently it
only affects the Ruby binding, but we plan to let the Ruby binding use
in-header forwarding states, too. The performance of VMs that use
in-header forwarding bits is not affected by this PR.

Fixes: https://github.com/mmtk/mmtk-core/issues/972