Stop using the name "ForwardingWord" (#976)
The name is likely inherited from JikesRVM. But since we changed its
name, we should just use the module name "object_forwarding" directly.

This cleans up some naming issues, and prepares for subsequent
refactoring that intends to change the logic of object forwarding in
ImmixSpace. At that time, I intend to fix this naming issue anyway.

Co-authored-by: Zixian Cai <2891235+caizixian@users.noreply.github.com>