Apply a delay before we enable auto merge on a PR (#992)
We saw the auto merging workflow (added in
https://github.com/mmtk/mmtk-core/pull/988) failed in
https://github.com/mmtk/mmtk-core/actions/runs/6587030083/job/17896533403.
The error message was `GraphQL: Head branch is out of date. Review and
try the merge again. (mergePullRequest)`. It is likely caused by the
issue https://github.com/orgs/community/discussions/24462: When a PR
gets updated, Github has a process to determine if the PR is mergable.
If we attempt to enable auto merge before the process is done, the merge
will fail with the error message above. This PR adds a delay and retry
for enabling auto merging.