Move some states from BasePlan to GlobalState (#949)
This PR moves some states from `BasePlan` to a new struct `GlobalState`.
Before this change, the states can only be accessed through `Plan`.
Whoever needs to access those states need to acquire or store a
reference to `Plan`, which is unnecessary. With the changes in the PR,
`GlobalState` is stored as `Arc<GlobalState>`. It can be directly
accessed through the `MMTK` instance, and can also be stored in some
components, such as allocators. This change reduces the dependency on
`Plan`.