Use `cargo generate-lockfile` to update JikesRVM's Cargo.lock (#996)
`auto-merge.yml` uses `cargo build` to generate lockfile for most
bindings. But this does not work for JikesRVM:
1. JikesRVM uses `--target i686-unknown-linux-gnu` which is not
installed automatically when running `cargo build`.
2. JikesRVM's cargo project uses some Rust source files that are
generated during building JikesRVM. So we cannot build the cargo project
alone from a fresh repo clone. We have to do a full build.

As a workaround, we use `cargo generate-lockfile` for JikesRVM. `cargo
generate-lockfile` will update the version of dependencies. But we do
not have a better option for JikesRVM