Post-release dependency version bump for v0.21.0
Bump the version of dependencies to their latest version.

The "atomic" crate, since version 0.6.0, requires the `T` in `Atomic<T>`
to implement `bytemuck::NoUnint`.  It is a marker trait for types that
satisfies some requirements.  One requirement is that `T` must not have
any padding bytes (in the middle or at the end).  The derive macro
`#[derive(NoUninit)]` can be used on custom types to automatically check
if the type satisfies its requirements.  Two notable types are:

1.  `enum MMapStrategy`.  It is missing a representation annotation
    which `NoUninit` requires.  This PR adds `#[repr(u8)]` to fix this.
2.  `WORKER_ORDINAL: Atomic<Option<ThreadId>>`.  The `Option` type does
    have padding bytes, making it un-eligible for `Atomic<T>`.  This PR
    changes it to `Atomic<ThreadId>`, with `ThreadId::max` representing
    the uninitialized value.