Revert a mistake in v0.22 that prevents cargo publish (#1065)
https://github.com/mmtk/mmtk-core/pull/1055 removed the version for
`mmtk-macros` in `Cargo.toml`. I thought it was fine, as it uses a local
path as the dependency. But `cargo publish` actually requires all the
dependencies to specify a version. So `cargo publish` failed for
https://github.com/mmtk/mmtk-core/releases/tag/v0.22.0 in
https://github.com/mmtk/mmtk-core/actions/runs/7284255107/job/19849376243.
The action did not return a non-zero code so I was not aware of the
issue. `v0.22` was not published to `crates.io`.

This PR adds back the version for `mmtk-macros`, adds a check for `cargo
publish` dry run in the style check, and fixes the publish script so the
script properly fails if all the attempts of `cargo publish` fail.