Resolve issues for unit tests on 32 bits (#1095)
This PR resolves a few issues we recently saw for running unit tests in
our CI.
* Fix an issue in the tests for byte map mmapper. In the clean up step,
we do munmap for a larger area than what we actually mmapped in the
test. This caused an issue that we may munmap regions that are used by
other code. https://github.com/mmtk/mmtk-core/issues/1092 is caused by
this -- the munmap after the tests unmaps the memory used by Rust, and
caused Rust to segfault after the tests.
* Change the test address for mmap tests to reduce the chance that we
may fail in mmap in the tests.
* ~~Make raw_memory_freelist only compiled in 64bits. This mitigates the
issue we saw in https://github.com/mmtk/mmtk-core/issues/1091 for
raw_memory_freelist tests.~~