Fix warnings on the stable branch.
As getting a ref `&` of `static mut` variables will become a compilation
error in the Rust 2024 Edition, we use `unsafe { &*addr_of!(VM_LAYOUT)
}` to work around it.

See: https://github.com/mmtk/mmtk-core/issues/1098

Changed the `static mut` in mock_test_doc_avoid_resolving_allocator.rs
to an ordinary local variable.

Initialize the thread-local variable `WORKER_ORDINAL` with a `const`
block.  It avoids lazy initialization.  We also bumped the
`rust-version` field in Cargo.toml to 1.71.1 because the semantics of
the `const` block in the `thread_local!` macro was undocumented in
1.70.0.