Rename associated types in GCWorkContext
This PR renames `ProcessEdgesWorkType` to `NormalProcessEdges`, and
renames `TPProcessEdges` to `PinningProcessEdges`.

The old name `TPProcessEdges` was very confusing because the type alone
is not sufficient to implement transitive pinning roots.  We rename it
to PinningProcessEdges, removing "transitive" from its name.  We also
renamed `ProcessEdgesWorkType` to `NormalProcessEdges` for symmetry.

FYI, non-transitive and transitive pinning roots are implemented using a
combination of `NormalProcessEdges` and `PinningProcessEdges`.

-   We use `PinningProcessEdges` for objects directly pointed from
    non-transitive pinning roots, but `NormalProcessEdges` for their
    descendents.
-   We use `PinningProcessEdges` to trace all objects reachable from
    transitive pinning roots.