Rename associated types in GCWorkContext (#1116)
This PR renames the associated types `ProcessEdgesWorkType` to
`DefaultProcessEdges`, and `TPProcessEdges` to `PinningProcessEdges`.

The old name `TPProcessEdges` was very confusing because that type alone
is not sufficient to implement transitive pinning roots. We rename it to
`PinningProcessEdges`, removing "transitive" from its name. We also
renamed `ProcessEdgesWorkType` to `DefaultProcessEdges` for symmetry. We
also updated comments to clarify the purposes of those type members.

We also updated the type parameters of
`ProcessEdgesWorkRootsWorkFactory<VM, DPE, PPE>` and
`ProcessRootNode<VM, R2OPE, O2OPE>` to clarify their purposes, at the
expense of being slightly more verbose.