Add missing pinning metadata to mock_vm (#1121)
If the "object_pinning" feature is enabled, we add the
`LOCAL_PINNING_BIT_SPEC` constant when implementing `ObjectModel` for
`MockVM`. That fixes compilation error if we attempt to run `cargo
clippy` with many features including both "mock_test" and
"object_pinning", although we currently don't do mock tests for the
pinning metadata.